backend/requirements.txt
# Backend Infrastructure Requirements for AI Crypto Trading Bot
# Version: 1.0
# Updated: 2025-10-07

# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Werkzeug==3.0.1

# Authentication & Security
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.1.2
cryptography==41.0.7

# Database
sqlite3  # Built-in with Python

# Data Processing
pandas==2.1.4
numpy==1.26.2

# HTTP Client
requests==2.31.0

# Scheduling & Background Tasks
APScheduler==3.10.4
schedule==1.2.1

# Logging & Monitoring
python-json-logger==2.0.7

# Date/Time Utilities
python-dateutil==2.8.2
pytz==2023.3

# Validation
jsonschema==4.20.0

# Rate Limiting Storage (Redis optional)
redis==5.0.1

# System Utilities
psutil==5.9.6

# Testing (optional)
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0

# Development Tools
black==23.12.1
flake8==7.0.0
mypy==1.7.1
